Use new mojular and npm
The new mojular framework works better if all assets are imported
as npm packages (and bower is being discontinued as of this writing).
This is a breaking change but won't affect applications using this
repo as it's imported by version/tag.
