+++ assets/scss/_base.scss
@@ -1,7 +1,3 @@
 .main-content {
   border-top: 1px solid transparent;
-
-  * + * {
-    margin-top: 1em;
-  }
 }
