--- bower.json
@@ -1,32 +1,0 @@
-{
-  "name": "money-to-prisoners-common",
-  "version": "0.5.4",
-  "homepage": "https://github.com/ministryofjustice/money-to-prisoners-common",
-  "authors": [
-    "Dom Smith <dom@teneightfive.com>"
-  ],
-  "main": [
-    "./assets/images/**",
-    "./assets/javascripts/**",
-    "./templates/**",
-    "./assets/scss/_mtp.scss"
-  ],
-  "description": "Common assets and template for the Money to Prisoners service",
-  "license": "MIT",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ],
-  "dependencies": {
-    "lodash": "~3.10.1",
-    "mojular": "ministryofjustice/mojular#0.7.0",
-    "hopscotch": "~0.2.5",
-    "js-cookie": "~2.0.3",
-    "jquery": "~1.11",
-    "hopscotch-highlight": "ministryofjustice/hopscotch-highlight#0.1.3",
-    "checked-polyfill": "rdebeasi/checked-polyfill"
-  }
-}
