+++ package.json
@@ -1,0 +1,48 @@
+{
+  "name": "money-to-prisoners-common",
+  "version": "0.5.3",
+  "description": "Assets common to all money-to-prisoners applications",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ministryofjustice/money-to-prisoners-common.git"
+  },
+  "author": "The Money To Prisoners team",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/ministryofjustice/money-to-prisoners-common/issues"
+  },
+  "homepage": "https://github.com/ministryofjustice/money-to-prisoners-common#readme",
+  "dependencies": {
+    "breakpoint-sass": "^2.6.1",
+    "browser-sync": "^2.10.0",
+    "chai": "^3.2.0",
+    "chai-as-promised": "^5.1.0",
+    "checked-polyfill": "ministryofjustice/checked-polyfill#1.0.0",
+    "cucumber": "^0.5.3",
+    "del": "^2.0.2",
+    "globule": "^0.2.0",
+    "gulp": "^3.9.0",
+    "hopscotch": "0.2.5",
+    "hopscotch-highlight": "ministryofjustice/hopscotch-highlight#1.0.0",
+    "imports-loader": "^0.6.5",
+    "jquery": "~1.11",
+    "js-cookie": "~2.0.3",
+    "mojular": "mojular/mojular#0.2.2",
+    "mojular-govuk-elements": "mojular/govuk-elements#0.2.3",
+    "mojular-moj-elements": "mojular/moj-elements#0.2.0",
+    "mojular-templates": "mojular/templates#0.0.8",
+    "money-to-prisoners-gulp-tasks": "ministryofjustice/money-to-prisoners-gulp-tasks#0.1.0",
+    "q": "^1.4.1",
+    "request": "^2.63.0",
+    "require-dir": "^0.3.0",
+    "sanitize-filename": "^1.4.3",
+    "selenium-standalone": "^4.6.0",
+    "webdriverio": "^3.2.4",
+    "webpack": "^1.12.3",
+    "yargs": "^3.24.0"
+  }
+}
