--- paths.json
@@ -1,5 +1,0 @@
-{
-  "import_paths": [
-    "money-to-prisoners-common/assets/scss"
-  ]
-}
