+++ templates/mtp_base.html
@@ -56,11 +56,11 @@
     </div>
   {% endif %}
 
-  <p class="subheader-item phase-banner">
-    <strong class="phase-tag phase-tag-alpha">{% trans "Alpha" %}</strong>
+  <div class="subheader-item phase-banner">
+    <div class="phase-tag phase-tag-alpha">{% trans "Alpha" %}</div>
     {# #TODO - add current location for feedback view #}
-    {% blocktrans %}This is a new service. If you have any comments or need help please <a href="#">contact us</a>.{% endblocktrans %}
-  </p>
+    <div class="phase-banner-message">{% blocktrans %}This is a new service. If you have any comments or need help please <a href="#">contact us</a>.{% endblocktrans %}</div>
+  </div>
 
   {% if messages %}
     <ul class="messages" role="alert" tabindex="-1">
